limit visibility, make stuff package private and refactor into util group
