+++ extensions/iceberg/src/main/java/io/deephaven/iceberg/util/IcebergTableAdapter.java
@@ -19,7 +19,6 @@ import io.deephaven.engine.table.impl.sources.regioned.RegionedTableComponentFac
 import io.deephaven.engine.util.TableTools;
 import io.deephaven.iceberg.base.IcebergUtils;
 import io.deephaven.iceberg.internal.DataInstructionsProviderLoader;
-import io.deephaven.iceberg.internal.Shim;
 import io.deephaven.iceberg.layout.IcebergAutoRefreshTableLocationProvider;
 import io.deephaven.iceberg.layout.IcebergBaseLayout;
 import io.deephaven.iceberg.layout.IcebergManualRefreshTableLocationProvider;
@@ -500,7 +499,7 @@ public final class IcebergTableAdapter {
         }
         final ParquetInstructions parquetInstructions = ParquetInstructions.builder()
                 .setTableDefinition(resolver.definition())
-                .setColumnResolverFactory(Shim.factory(resolver, nameMapping))
+                .setColumnResolverFactory(new ResolverFactory(resolver, nameMapping))
                 .setSpecialInstructions(specialInstructions)
                 .build();
         final Map<String, PartitionField> partitionFields = resolver.partitionFieldMap();
