+++ extensions/iceberg/src/main/java/io/deephaven/iceberg/util/InferenceImpl.java
@@ -7,7 +7,6 @@ import io.deephaven.api.util.NameValidator;
 import io.deephaven.base.verify.Assert;
 import io.deephaven.engine.table.ColumnDefinition;
 import io.deephaven.engine.table.TableDefinition;
-import io.deephaven.iceberg.internal.PartitionSpecHelper;
 import io.deephaven.qst.type.Type;
 import org.apache.iceberg.PartitionField;
 import org.apache.iceberg.Schema;
