+++ extensions/iceberg/src/test/java/io/deephaven/iceberg/util/ResolverFactorySimpleTest.java
@@ -1,12 +1,10 @@
 //
 // Copyright (c) 2016-2025 Deephaven Data Labs and Patent Pending
 //
-package io.deephaven.iceberg.internal;
+package io.deephaven.iceberg.util;
 
 import io.deephaven.engine.table.ColumnDefinition;
 import io.deephaven.engine.table.TableDefinition;
-import io.deephaven.iceberg.util.ColumnInstructions;
-import io.deephaven.iceberg.util.Resolver;
 import io.deephaven.parquet.table.location.ParquetColumnResolver;
 import org.apache.iceberg.Schema;
 import org.apache.iceberg.mapping.MappedField;
