+++ extensions/iceberg/src/test/java/io/deephaven/iceberg/util/SchemaHelperTest.java
@@ -1,7 +1,7 @@
 //
 // Copyright (c) 2016-2025 Deephaven Data Labs and Patent Pending
 //
-package io.deephaven.iceberg.internal;
+package io.deephaven.iceberg.util;
 
 import org.apache.iceberg.Schema;
 import org.apache.iceberg.types.Types.IntegerType;
